Tweaks and fixes in advance of integration:   * cache volume capacity needs to be given as a long   * BySizeSelectionStrategy: ensure positive normalizing paramter   * make grace period configurable

+++ src/main/java/gov/nist/oar/distrib/cachemgr/BasicCache.java
@@ -262,7 +262,7 @@ public abstract class BasicCache extends Cache {
      * @throws InventoryException  if a problem occurs while registering the volume with 
      *                  the inventory database.
      */
-    public void addCacheVolume(CacheVolume vol, int capacity, JSONObject metadata, boolean updmd)
+    public void addCacheVolume(CacheVolume vol, long capacity, JSONObject metadata, boolean updmd)
         throws CacheManagementException
     {
         if (metadata == null)
